requirements.txt
numpy==1.24.1
pip==21.3.1
requests==2.26.0
psycopg2-binary==2.9.3
pandas==1.3.0